Duke Spy costume optimization
"spycloak00 is 174360 bytes because it is IA8. However, the alpha is a uniform setting throughout.

Since a material is needed anyway to use it, you can optimize out OVER HALF of this by converting it to I8 and simply going into the material, changing Color0 Alpha RGBA 255,255,255,0 to 255,255,255,188, then going into the shader stage's Alpha Output and changing Selection B from TextureAlpha to Alpha0."